ENH: Added pyproj.database.query_crs_info
