+++ docs/api/database.rst
@@ -1,0 +1,38 @@
+Database
+=========
+
+Methods that query the PROJ database for information.
+
+
+pyproj.database.get_units_map
+-----------------------------
+
+.. note:: The backwards compatible import is `pyproj.get_units_map`
+
+.. autofunction:: pyproj.database.get_units_map
+
+.. autoclass:: pyproj.database.Unit
+
+
+pyproj.database.get_authorities
+--------------------------------
+
+.. note:: The backwards compatible import is `pyproj.get_authorities`
+
+.. autofunction:: pyproj.database.get_authorities
+
+
+pyproj.database.get_codes
+--------------------------
+
+.. note:: The backwards compatible import is `pyproj.get_codes`
+
+.. autofunction:: pyproj.database.get_codes
+
+
+pyproj.database.query_crs_info
+-------------------------------
+
+.. autofunction:: pyproj.database.query_crs_info
+
+.. autoclass:: pyproj.database.CRSInfo
