+++ docs/api/sync.rst
@@ -7,12 +7,6 @@ pyproj.sync.get_transform_grid_list
 .. autofunction:: pyproj.sync.get_transform_grid_list
 
 
-pyproj.sync.BBox
------------------
-
-.. autoclass:: pyproj.sync.BBox
-    :members:
-
 pyproj.sync.get_proj_endpoint
 ------------------------------
