+++ pyproj/_compat.pxd
@@ -1,0 +1,4 @@
+from pyproj.compat import pystrdecode
+
+
+cdef cstrdecode(const char *instring)
