+++ pyproj/list.pyi
@@ -1,0 +1,5 @@
+from typing import Dict
+
+def get_proj_operations_map() -> Dict[str, str]: ...
+def get_ellps_map() -> Dict[str, Dict[str, float]]: ...
+def get_prime_meridians_map() -> Dict[str, str]: ...
